npm test => make test
Since there is a make file with config for testing, it makes sense to specify this in `package.json`.

The `Makefile` sets the mocha spec reporter, which I agree is much more readable than the dots. Still not sure why dots is default?!? Might as well take advantage of the better report for `npm test`.

Might also want to update the README accordingly?